Removed unecessary casts.

+++ xpilot/src/common/shipshape.c
@@ -235,7 +235,7 @@ static int shape2wire(char *ship_shape_str, shipshape_t *ship)
 	str++;
 
 	if (shape_version == 0) {
-	    if (isdigit((int)*str)) {
+	    if (isdigit(*str)) {
 		shape_version = 0x3100;
 		if (verboseShapeParsing)
 		    xpprintf("ship shape is in old format\n");
@@ -1369,14 +1369,14 @@ static int Get_shape_keyword(char *keyw)
     int			i;
 
     /* non-abbreviated keywords start with a lower case letter. */
-    if (islower((int)*keyw)) {
+    if (islower(*keyw)) {
 	for (i = 0; i < NUM_SHAPE_KEYS; i++) {
 	    if (!strcmp(keyw, shape_keys[i]))
 		break;
 	}
     }
     /* abbreviated keywords start with an upper case letter. */
-    else if (isupper((int)*keyw)) {
+    else if (isupper(*keyw)) {
 	for (i = 0; i < NUM_SHAPE_KEYS; i++) {
 	    if (!strcmp(keyw, abbrev_keys[i]))
 		break;
